# perf: use argon2 in node to speedup cli account decoding