--- a/src/typings/@aeternity__argon2-browser/index.d.ts
+++ b/src/typings/@aeternity__argon2-browser/index.d.ts
@@ -1,3 +1,0 @@
-declare module '@aeternity/argon2-browser/dist/argon2-bundled.min.js' {
-  export * from 'argon2-browser';
-}