Keystore tests before:
```
Keystore
    ✔️ dump account to keystore object (608ms)
    ✔️ dump accepts hex (504ms)
    ✔️ restore account from keystore object (498ms)
    ✔️ use invalid keystore password (498ms)
...
84 passing (2s)
```
Keystore tests after:
```
Keystore
    ✔️ dump account to keystore object (165ms)
    ✔️ dump accepts hex (130ms)
    ✔️ restore account from keystore object (130ms)
    ✔️ use invalid keystore password (132ms)
...
84 passing (808ms)
```